requirements.txt
python-pptx
pywin32
opencv-python
pytesseract
pil
tesseract
tesseract-ocr
natsort